requirements.txt
attrs==15.2.0
beautifulsoup4==4.4.1
blinker==1.3
cffi==1.6.0
chardet==2.3.0
chicken==0.1.0
cryptography==1.2.3
Django==1.9.6
django-ajax-toolkit==0.3.0
django-autoslug==1.9.3
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.10.0
django-rgbfield==1.0
django-social-auth==0.7.28
django-static-precompiler==1.4
django-taggit==0.18.3
django-taggit-autosuggest==0.2.8
django-tinymce==2.3.0
django-voting==0.2
egg==0.2.0
enum34==1.1.2
html5lib==0.999
httplib2==0.9.2
idna==2.0
ipaddress==1.0.16
lxml==3.5.0
msgpack-python==0.3.0
oauth2==1.9.0.post1
oauthlib==1.0.3
pbr==1.10.0
Pillow==3.2.0
piston-mini-client==0.7.5
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycparser==2.14
pycups==1.9.73
PyJWT==1.4.0
pyOpenSSL==0.15.1
pyserial==3.0.1
python-openid==2.2.5
python-social-auth==0.2.19
pyxdg==0.25
requests==2.10.0
requests-oauthlib==0.6.1
service-identity==16.0.0
six==1.10.0
stevedore==1.14.0
Twisted==16.0.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
zope.interface==4.1.3

South==1.0
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
